Fixed an issue where an out of bounds exception would be thrown if a mod had more than 127 masters

+++ src/skyproc/FormID.java
@@ -192,9 +192,9 @@ public class FormID implements Comparable, Serializable {
 
     private void adjustMasterIndex(Mod srcMod) {
 	ArrayList<String> masters = srcMod.getMastersStrings();
-	for (byte i = 0; i < masters.size(); i++) {
+	for (int i = 0; i < masters.size(); i++) {
 	    if (masters.get(i).equals(master.print())) {
-		form[3] = i;
+		form[3] = (byte) i;
 		return;
 	    }
 	}
